requirements.txt
streamlit
pillow
requests
google-generativeai
discogs-client
spotipy
beautifulsoup4
lxml
aiohttp
fuzzywuzzy
python-Levenshtein
pytesseract